requirements.txt
openai==0.27.0
pytest==7.2.2
python-dotenv==0.19.1
